gcs-function/requirements.txt
functions-framework>=3.8.3      # CloudEvent 핸들러용
cloudevents>=1.10.1
google-cloud-storage>=2.16.0
pandas>=2.2.2
google-cloud-secret-manager
google-cloud-discoveryengine>=0.9.0 
# Excel
openpyxl>=3.1.2

# parsing libs
pdfplumber>=0.11.0
olefile>=0.46           # hwp·OLE 파일 열기용
pymupdf>=1.24.3  

# openai 등을 parsers.py 에서 사용한다면
openai>=1.30.1
